+++ flick_urban/geo4cfd/ansa/DividePrecursor.py
@@ -1,5 +1,8 @@
 # PYTHON script
 import os
+import sys
+sys.path.insert(0, os.path.dirname(os.path.abspath(__file__)))
+
 import ansa
 from RebuildInlet import DividePIDbyNormalVector
 from ansa import *